Switching out *_release tables for *_transaction tables
